# parser.rl: parse_string implement a fast path