@casperisfine 2024-11-06
Not too sure why it helps quite a lot on `activitypub`, but less on `twitter` and almost not on `ctim_catalog`.