@casperisfine 2024-11-06
> Not too sure why it helps quite a lot on `activitypub`, but less on `twitter` and almost not on `ctim_catalog`.

Alright, after instrumenting:

`citm_catalog.json`:

  - cache_hit:  25576
  - fastpath: 1027
  - slowpath: 1

`twitter.json`:

  - cache_hit: 13345
  - fastpath: 4442
  - slowpath:  312

`activitypub.json`:

  - cache_hit: 878
  - fastpath: 598
  - slowpath: 59

So that explain why they don't all benefit the same from this.